changed MealFragment and InputMealFOrm to use the database (Still need to add in the barcode scanner, search feature, and delete functionality)

+++ app/src/main/java/com/example/nutritiontrackerguiv4/database/IngredientDAO.java
@@ -23,7 +23,7 @@ public interface IngredientDAO {
     List<Ingredient> getAllIngredients();
 
     @Query("SELECT * FROM Ingredients WHERE Ingredient_ID=:ingID")
-    List<Ingredient> findAllInfoForIngredient(final int ingID);
+    List<Ingredient> findAllInfoForIngredient(final long ingID);
 
 
 }
